Move to push-parse for H264 NAL Unit syntax
Buffering the data and then doing a pull-parse once the entire PES packet was
visible is quite costy becase of all the data we copied around in memory.

The hand-rolled parser state machine in this new model is vastly more complicated
but also makes big reduction in the cost of one of the most expensive code-paths.