Optimise: don't parse all H264 data all the time
We only switch on H264 parsing periodically to check that our calculated
relationship between PCR values and SEI pic_timing values is still correct.

H264 parsing is a relatively expensive part of the protocol stack, so this
makes a fair difference to the overall cost when chunking higher bitrate
video.