Optimise writes by gathering multiple packets.
We previously wrote one transport stream packet at a time to disk.
By collecting several sequential packets into a CompositeByteBuf,
we are not actually copying data in memory, but we are able to
send all the pieces to the kernel in one gathering write operation
(e.g. writev() on linux).

This change takes processing time for a benchmark file I tested with
from ~13 seconds to ~6 seconds.