Enable content compression; check Content-Length
I suspect some intermediaries  on the internet will include

  Accept-Encoding: gzip

in requests to origin, and then decompress any compressed response if we did
not send it ourselves.  A side effect of this is that they will not send us
a Content-Length up front, even if the origin included Content-Length (for
the original compressed size) in its response.

By announcing support for content compression, we hopefully see the origin's
response, and can assert Content-Length was present (although we need to
workaround the Apache HttpClient ResponseContentEncoding filter removing
Content-Length before application code gets to see the response.