MediaContext becomes a factory for Locator objects
Previously, Locator objects where passed around within the representations
of units of media (e.g. as an attribute of a Transport Stream packet) or
on the stack as arguments to method calls.  This meant the creation of many
Locator instances even when the where never subsequently used.  It also
complicated lots of APIs, and was not consistent.  Now, with a chain of
MediaContext objects available for the full media processing stack, the
contexts can have all the information required to construct a Locator on
the fly, and provide a much more understandable interface for their
acquisition.